nodes and relations are updated

+++ recommendation/src/main/java/com/stackroute/juggler/recommendation/repositories/CityRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.neo4j.repository.Neo4jRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.stackroute.juggler.kafka.domain.City;
+import com.stackroute.juggler.recommendation.domain.City;
 @Repository
-public interface CityRepository extends Neo4jRepository<City, Integer>{
+public interface CityRepository extends Neo4jRepository<City, Long>{
 
 	
 	@Query("MATCH (c:City) WHERE c.name ={name} RETURN c")
